ejercicio para entregar Blade of Darkness me falta el metodo golpear del monstruo y del jugador casi hecho

+++ CursoJava2025/SegundaEvaluacion/Tema4/EjercicioEntregar/BladeOfDarkness/Arma.java
@@ -29,7 +29,7 @@ public class Arma {
         this.tipo = tipo;
     }
 
-    public Integer getPuntosD() {
+    public Double getPuntosD() {
         return puntosD;
     }
